Initial script for calculating dram timings for Cubietruck
It takes the desired dram clock frequency as a command line argument
and produces a hopefully optimal 'dram_para' struct for u-boot.

Kudos to Jens Kuske (jemk) for discovering the 'magic' tpr3 value,
which is a real game changer:
    http://irclog.whitequark.org/linux-sunxi/2014-04-07#7042945
    http://irclog.whitequark.org/linux-sunxi/2014-04-08#7048651

The 528MHz dram clock frequency is really interesting, because
it is close to the "standard" 32-bit DDR3-1066. And we may hope
that it may turn out ot be reliable on all Cubietrucks.

But it is expected that many Cubietrucks can actually run dram at
a much higher clock frequency than that.