'scan-for-best-tpr3.rb' is now launched by a10-watchdog
So we don't need to launch a10-watchdog from this script itself
anymore. Also we don't issue a "reboot" command at the end, but
return an exit code for the external a10-watchdog tool. This works
a bit nicer for testing purposes when we don't particularly want
to reboot the system after this script is done.