A partial rewrite of scan-for-best-tpr3.rb/parse-tpr3-results.rb
Now we collect information about the host name and dcdc3 voltage
in order to more correctly distinguish different configurations.

Memtester logs are scanned to identify the problematic byte lanes
and this information is visualized in the report html as bold
digits of the tpr3 value corresponding to the failing lane.

Some statistics is calculated in the hope that it is useful.

Now after all the basic job is done, we are trying to run an
additional round of more extensive tests for the 'GREEN' results.

But still a lot has to be done and a major cleanup is needed.