scan-for-best-tpr3.rb: Rely less on bitspread and stuck address tests
The bitspread memtester test is to be avoided because it is not
using alternating v and ~v values to fill the memory. So it is not
very good for detecting and reporting errors in multiple lanes
simultaneously. Still the bitspread test is able to detect some
problems, which are difficult to reproduce with the other tests.
So we still use it, but only later on.

The stuck address test is just a waste of time and also it is not
suitable for detecting problematic lanes.