scan-for-best-tpr3.rb: Optimize the order of the tpr3 values scan
Return a sequence of tpr3 values, prioritizing the ones which are more
likely to be "good". The good ones are assumed to be near other good
results. And in the case of a tie, prioritize the tpr3 values, which
are close to the the original tpr3 setup (configured by the bootloader).

Basically, the scan is now optimized for locating large "green"
clusters as early as possible.