scan-for-best-tpr3: Added a SoC cool down pause
Try to stress a wider range of temperatures during the test. For this
we just sleep for 100 seconds after the second iteration (running
two iterations before sleeping is used in order not to waste time
on obviously broken configurations). During the sleep pause, the
temperature of the SoC drops. And after the memtester runs resume,
the temperature starts climbing up again.