A 'luminance' based model for calculating the stability scores
We treat each cell in the table as a light emittter (the intensity
is proportional to the number of successful memtester runs). And
then try to find the maximum luminance at different heights. For
lower heights, the contribution of the immediate neighbours is
much higher. For larger heights, the contribution of different
table cells is treated more equally across a larger area.

Also updated the boilerplate text to provide a new link to the
github project page and to the linux-sunxi wiki.