a10-tpr3-common.rb: Prioritize 3x3 area around the default tpr3
Change the sorting order handling so that the 3x3 area around
the default tpr3 value (the value, which is configured by U-Boot)
is always handled first. This is especially important for running
the final hardening tests, because this way they can be used to
esimate whether the current U-Boot DRAM setup is good.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>